Add Ui.java class which deals with interactions with the user
